backend/requirements.txt
Flask
mysql-connector-python
python-dotenv
flask-jwt-extended
flask-cors